ai-worker/requirements.txt
gradio
insightface
onnxruntime
opencv-python
Pillow
numpy
